new logic

+++ app/Http/Controllers/OptionController.php
@@ -2,7 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Contact;
 use App\Models\Option;
+use App\Models\Student;
+use App\Models\Transaction;
 use Illuminate\Http\Request;
 
 class OptionController extends Controller
@@ -14,14 +17,32 @@ class OptionController extends Controller
         $debit = Option::where('type', Option::TYPE_DEBIT)->get();
         $school = Option::where('type', Option::TYPE_SCHOOL)->get();
         $wallet = Option::where('type', Option::TYPE_WALLET)->get();
+        $studentStatus = Option::where('type', Option::TYPE_STUDENT_STATUS)->get();
+        $donaturStatus = Option::where('type', Option::TYPE_DONATUR_STATUS)->get();
+        $employeeStatus = Option::where('type', Option::TYPE_EMPLOYEE_STATUS)->get();
+        $employeeCategory = Option::where('type', Option::TYPE_EMPLOYEE_CATEGORY)->get();
         return view('admin.options.index', [
             'pck' => Option::TYPE_PACKAGE,
             'crd' => Option::TYPE_CREDIT,
             'dbt' => Option::TYPE_DEBIT,
             'sch' => Option::TYPE_SCHOOL,
             'cls' => Option::TYPE_CLASS,
-            'wlt' => Option::TYPE_WALLET
-        ], compact('donatePackage', 'class', 'credit', 'debit', 'school', 'wallet'));
+            'wlt' => Option::TYPE_WALLET,
+            'std_sts' => Option::TYPE_STUDENT_STATUS,
+            'dnt_sts' => Option::TYPE_DONATUR_STATUS,
+            'emp_sts' => Option::TYPE_EMPLOYEE_STATUS,
+            'emp_ctg' => Option::TYPE_EMPLOYEE_CATEGORY,
+        ], compact(
+            'donatePackage',
+            'class',
+            'credit',
+            'debit',
+            'school',
+            'wallet',
+            'studentStatus',
+            'donaturStatus',
+            'employeeStatus',
+            'employeeCategory'));
     }
 
     public function submit (Request $request) {
@@ -71,14 +92,32 @@ class OptionController extends Controller
         return redirect()->back()->with('success', 'Data Berhasil Disimpan');       
     }
 
-    public function delete ($id) {
+    public function delete($id) {
         $options = Option::find($id);
+    
+        $status = Contact::where('status_id', $id)->exists();
+        $schoolUsed = Contact::where('school_id', $id)->exists();
+        $classUsed = Contact::where('class_id', $id)->exists();
+        $donatePackage = Contact::where('package_id', $id)->exists();
+        $wallet = Transaction::where('wallet_id', $id)->exists();
+        $trx_category = Transaction::where('payable_id', $id)->exists();
+        $employeeCategory = Contact::where('employee_category_id')->exists();
+    
+        if ($status || $schoolUsed || $classUsed || $donatePackage || $wallet || $trx_category || $employeeCategory) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Data yang sudah digunakan, tidak bisa dihapus!',
+                'redirect' => route('options.index')
+            ]);
+        }
+    
         $options->delete();
-
+    
         return response()->json([
             'success' => true,
-            'message' => 'Data berhasil dihapus',
+            'message' => 'Data berhasil dihapus!',
             'redirect' => route('options.index')
         ]);
     }
+    
 }
